refactor: extract client-metadata to shared package with CSS branding support
Move resolveClientMetadata, resolveClientName, and ClientMetadata type
from auth-service into @certified-app/shared so both pds-core and
auth-service can use them. Add branding.css to ClientMetadata interface,
escapeCss() for safe HTML embedding, and getClientCss() trust-gated helper.

Auth-service's client-metadata.ts becomes a re-export shim.

Ref: bd-20x.1
